Add DICOM metadata extraction and use in volume estimation
Implemented reading of DICOM metadata (pixel spacing, slice thickness, etc.) and integrated it into MRI sequence construction and cancer volume estimation. Updated CancerVolume to use Float for volume. Metadata is now extracted from the first DICOM file and used throughout the app for more accurate volume calculations.

+++ app/src/main/java/com/example/demoapp/Model/CancerVolume.kt
@@ -1,7 +1,7 @@
 package com.example.demoapp.Model
 
 class CancerVolume (
-    var volume : Int,
+    var volume : Float,
     var sequence: MRISequence,
     var affinityMatrix: Array<Array<FloatArray>>,
 )
